Binding Start Touch Bar action to TimerBar. Refactor the name of TouchBarService (mac)
