+++ src/ui/osx/TogglDesktop/Feature/TouchBar/TouchBarService+Name.swift
@@ -1,0 +1,22 @@
+//
+//  TouchBarService+Name.swift
+//  TogglDesktop
+//
+//  Created by Nghia Tran on 10/8/19.
+//  Copyright © 2019 Alari. All rights reserved.
+//
+
+import Foundation
+
+@available(OSX 10.12.2, *)
+extension NSTouchBar.CustomizationIdentifier {
+    static let timeEntry = NSTouchBar.CustomizationIdentifier("com.toggl.toggldesktop.timeentrytouchbar")
+}
+
+@available(OSX 10.12.2, *)
+extension NSTouchBarItem.Identifier {
+
+    static let timeEntryItem = NSTouchBarItem.Identifier("com.toggl.toggldesktop.timeentrytouchbar.timeentryitems")
+    static let runningTimeEntry = NSTouchBarItem.Identifier("com.toggl.toggldesktop.timeentrytouchbar.runningtimeentry")
+    static let startStopItem = NSTouchBarItem.Identifier("com.toggl.toggldesktop.timeentrytouchbar.startstopbutton")
+}
